Fixes code refactoring , basket counter

+++ ArtificialOrdersInput/CreateOrders.cs
@@ -39,7 +39,7 @@ namespace ArtificialOrdersInput
                {
                   Guid productId = Guid.Parse(order.GetProperty("id").GetString());
                   ushort amount = order.GetProperty("amount").GetUInt16();
-                  await _orderListService.AddOrderAsync((Guid)userId, productId, amount);
+                  await _orderListService.AddOrderAsync((Guid)userId, productId, (short)amount);
                }
             await _orderListService.SubmitOrder((Guid)userId);
